weather-BOT/requirements.txt
flask==2.1.2
line-bot-sdk
requests
